feat: [prd-embedded-observer-graph-editor-005] - [Retire active-flow dependency on the separate editor graph and keep interactions stable]
