+++ docs/internal/processes/development-guide-relevant-files.md
@@ -258,8 +258,9 @@ This inventory records the checked-in files and directories that the maintainer 
 - When factory-graph editor layout work needs observer parity, route editor node positioning through `ui/src/features/factory-graph-editor/factory-graph-editor-layout.ts` and the shared layered-layout helper in `ui/src/features/flowchart/layout.ts` instead of reintroducing fixed kind-column placement inside `factory-graph-editor-flow.tsx`.
 - When multiple workflow-activity graph consumers need topology-keyed async layout reuse, keep the cache and in-flight coordination in `ui/src/features/workflow-activity/hooks/workflow-topology-async-cache.ts` and let each caller keep its own topology key, layout builder, and result shaping so observer and editor flows share concurrency behavior without growing an app-wide cache subsystem.
 - When `ui/src/features/factory-graph-editor/factory-graph-draft-apply.ts` empties an optional workstation route or resource field, delete the property instead of storing explicit `undefined` so pending factory definitions stay in canonical API shape and equality checks match serialized output.
-- When factory-graph editor visibility changes are presentation-only, keep the preset or filter state in `ui/src/features/workflow-activity/hooks/react-flow-current-activity-card-editor-graph.ts` and filter the rendered topology there so switching views never mutates or discards pending draft graph edits.
+- When embedded workflow-activity graph editing grows new editor-only affordances, layer them onto the shared observer seams in `ui/src/features/workflow-activity/components/react-flow-current-activity-card.tsx`, `.../react-flow-current-activity-card-surface.tsx`, and `.../react-flow-current-activity-card-viewport.tsx` instead of reviving a second editor-only graph view-model; presentation-only editor state should not mutate or fork the saved observer topology.
 - When Storybook responsive verifiers target graph-editor toolbar or preset controls, anchor Playwright locators to the component's stable accessible label or labeled region and use exact accessible-name matching for the child controls so implicit-role differences or overlapping node-handle names do not trip strict-mode failures.
+- When Storybook responsive verifiers target graph-editor toolbar or preset controls, scope Playwright role queries to the owning labeled region and use exact accessible-name matching so node-handle controls with overlapping substrings cannot trip strict-mode locator failures.
 - When a repo-owned Storybook verifier switches from one Playwright query shape to another, keep the focused Vitest fake page in the sibling `ui/scripts/*.test.mjs` file aligned with the same page methods (`locator`, `getByRole`, `getByTitle`, and similar) so CI does not fail on a test-double API mismatch that the live browser script itself would not hit.
 - When backend size cleanup touches oversized `pkg/config` test owners such as `config_mapper_test.go` or `runtime_config_test.go`, split them into same-package sibling files by authored-mapping, runtime-load, named-factory, or workstation-kind concern, and keep shared JSON fixture writers plus lookup/assert helpers in dedicated package-local `*_test_helpers_test.go` files instead of duplicating them across the new test owners.
 - When backend size cleanup touches `pkg/config` boundary-heavy test owners such as `config_validator_test.go`, `factory_config_mapping_test.go`, or `openapi_factory_test.go`, split semantic-validator, generated-boundary, and public-OpenAPI assertions into same-package sibling files while keeping shared finding, payload, and alias-assertion helpers package-local instead of scattering them across unrelated packages.
