+++ docs/internal/processes/factory-workstation-relevant-files.md
@@ -88,7 +88,7 @@ work inputs.
 - When a workstation edit surface exposes worker-backed fields such as `model`, verify whether the referenced worker is shared by sibling workstations before treating the field as workstation-scoped; if workers can be shared, disable or clearly rescope that edit so one workstation save cannot silently rewrite others.
 - When a dirty workstation draft survives a newer current-factory refresh, keep the session-start draft and the latest server-backed draft side by side; derive overwrite warnings by comparing the current draft to the latest server draft instead of rehydrating away the customer's pending input.
 - When `FACTORY_CHANGE` events can arrive while the graph editor is open, update the normalized current-factory cache and invalidate the current-factory document query together; stale-save warnings depend on refreshed version metadata, not just refreshed topology payloads.
-- Keep explicit graph-editor save/discard chrome in dedicated editor modules such as `components/react-flow-current-activity-card-editor-dialogs.tsx`, `components/react-flow-current-activity-card-editor-chrome.tsx`, and `components/factory-graph-editor-draft-actions.tsx`; that keeps confirmation, stale-warning, and dirty-state behavior reusable without growing the main dashboard card past lint thresholds.
+- Keep explicit graph-editor save/discard chrome in dedicated editor modules such as `components/react-flow-current-activity-card-editor-dialogs.tsx`, `components/react-flow-current-activity-card-editor-chrome.tsx`, and `ui/src/features/factory-graph-editor/components/factory-graph-editor-controls.tsx`; that keeps confirmation, stale-warning, and dirty-state behavior reusable without growing the main dashboard card past lint thresholds.
 - For coverage-sensitive graph-editor fixes, prefer focused tests at the small controller, dialog, and copy seams (`react-flow-current-activity-card-editor-controllers.test.tsx`, `components/factory-graph-editor-add-dialog.test.tsx`, and edge-removal copy tests), and keep `components/react-flow-current-activity-card.coverage.test.tsx` limited to viewport wiring that needs the React Flow shell mocked.
 - When browser replay or export integration fixtures stand up a lightweight dashboard API server, mirror the canonical current-factory editing contract there too: `GET` and `PUT /factory-sessions/{session_id}/factory` need the canonical `Factory` payload shape plus CORS headers, or graph-editor fetches will fail in replay lanes even when the production dashboard code is correct.
 - When a selection-scoped dashboard mutation needs prominent affordance without changing the page toolbar contract, prefer the existing `SelectionDetailLayout` `headerAction` slot; it keeps feature-specific primary actions beside shared undo/redo controls while leaving the global dashboard header focused on page-level controls.
@@ -164,7 +164,7 @@ work inputs.
 - Three-workstation chaining-lineage smoke coverage can stay fully behavioral by scaffolding the factory in-test, using one worker-emitted `FACTORY_REQUEST_BATCH` fan-out to create branch work with source-dispatch tags, and asserting the mixed work-plus-resource fan-in only through canonical events, world-state reconstruction, and final token state.
 - Graph-editor connection UX should keep anchor compatibility ownership in `ui/src/features/factory-graph-editor/factory-graph-editor-connections.ts`, render labeled anchor seams through `components/factory-graph-editor-flow.tsx`, and verify keyboard-reachable anchor states through the focused flow Storybook story instead of the broader dashboard surface.
 - Graph-editor edge deletion should keep draft edge add/remove mutations in `ui/src/features/factory-graph-editor/factory-graph-editor-connections.ts`, keep removal confirmation copy in `factory-graph-editor-removals.ts`, and use the focused `components/factory-graph-editor-flow.stories.tsx` seam for browser-backed pending-edge styling proof instead of broader React Flow card tests.
-- Graph-editor worker runtime badges and dense-lane toggles should stay derived at the editor view-model layer (`ui/src/features/workflow-activity/hooks/react-flow-current-activity-card-editor-graph.ts`) from the canonical editable definition plus live dashboard runtime data; keep badge styling and focused browser verification in `ui/src/features/factory-graph-editor/components/factory-graph-editor-flow.tsx` and `components/factory-graph-editor-flow.stories.tsx` instead of rebuilding worker status logic inside the broader dashboard card.
+- Retired editor-only graph lanes should stay isolated to the standalone factory-graph editor seams under `ui/src/features/factory-graph-editor/components/factory-graph-editor-flow.tsx`; the embedded workflow-activity editor should keep using the shared observer graph surface instead of reintroducing worker/resource lane rendering or a second runtime graph model.
 - Graph-editor draft transformation helpers should keep `ui/src/features/factory-graph-editor/factory-graph-draft-apply.ts` as the canonical owner for named-entity, work-state, pending-definition, and workstation worker-assignment application; validation can reuse that owner, but edge-removal validation must still seed removed endpoints into its node index so deleting a node and its incident edge in one draft does not regress the existing save/validation behavior.
 - Graph-editor pending-save and validation regressions around workstation worker assignments should be proved in the focused draft behavioral seams under `ui/src/features/factory-graph-editor/lib/`, with replacement and removal-without-replacement coverage in `factory-graph-draft.state.test.tsx` and shared fixtures in `factory-graph-draft.test-helpers.ts`, instead of helper-location or UI-copy-only tests.
 - Graph-editor draft validation coverage for duplicate identifiers, required-name failures, incompatible edges, unknown edge nodes, and projection fallback should stay in the focused `ui/src/features/factory-graph-editor/lib/factory-graph-draft.definition.test.tsx` behavioral seam, asserting returned validation errors and `buildPendingFactoryDefinition(...)` nullability rather than private validator helpers.
