+++ ui/scripts/hardcoded-ui-copy-baseline.txt
@@ -148,17 +148,17 @@ src/features/factory-graph-editor/lib/factory-graph-editor-save-summary.ts|48|12
 src/features/factory-graph-editor/lib/factory-graph-editor-save-summary.ts|52|12|string-literal|This save will apply .
 src/features/factory-graph-editor/lib/factory-graph-editor-save-summary.ts|56|10|string-literal|This save will apply and .
 src/features/flowchart/components/current-activity-node-chrome.tsx|26|10|string-literal|text-[0.65rem] font-semibold uppercase tracking-[0.08em]
-src/features/flowchart/components/current-activity-place-node.tsx|151|14|string-literal|border-af-overlay/22 bg-af-canvas text-af-ink
-src/features/flowchart/components/current-activity-place-node.tsx|153|12|string-literal|border-dashed border-af-info/36 bg-af-surface/78 text-af-ink
-src/features/flowchart/components/current-activity-place-node.tsx|168|14|string-literal|Terminal
-src/features/flowchart/components/current-activity-place-node.tsx|171|14|string-literal|Failed
-src/features/flowchart/components/current-activity-place-node.tsx|173|12|string-literal|Queue
-src/features/flowchart/components/current-activity-place-node.tsx|177|12|string-literal|Resource
-src/features/flowchart/components/current-activity-place-node.tsx|180|35|string-literal|Limit
-src/features/flowchart/components/current-activity-place-node.tsx|180|45|string-literal|Constraint
-src/features/flowchart/components/current-activity-place-node.tsx|208|12|string-literal|Processing state
-src/features/flowchart/components/current-activity-place-node.tsx|281|12|string-literal|resource tokens
-src/features/flowchart/components/current-activity-workstation-node.tsx|227|21|jsx-prop|Active
+src/features/flowchart/components/current-activity-place-node.tsx|156|14|string-literal|border-af-overlay/22 bg-af-canvas text-af-ink
+src/features/flowchart/components/current-activity-place-node.tsx|158|12|string-literal|border-dashed border-af-info/36 bg-af-surface/78 text-af-ink
+src/features/flowchart/components/current-activity-place-node.tsx|173|14|string-literal|Terminal
+src/features/flowchart/components/current-activity-place-node.tsx|176|14|string-literal|Failed
+src/features/flowchart/components/current-activity-place-node.tsx|178|12|string-literal|Queue
+src/features/flowchart/components/current-activity-place-node.tsx|182|12|string-literal|Resource
+src/features/flowchart/components/current-activity-place-node.tsx|185|35|string-literal|Limit
+src/features/flowchart/components/current-activity-place-node.tsx|185|45|string-literal|Constraint
+src/features/flowchart/components/current-activity-place-node.tsx|213|12|string-literal|Processing state
+src/features/flowchart/components/current-activity-place-node.tsx|286|12|string-literal|resource tokens
+src/features/flowchart/components/current-activity-workstation-node.tsx|231|21|jsx-prop|Active
 src/features/flowchart/components/graph-semantic-icon.tsx|39|12|string-literal|Active work
 src/features/flowchart/components/graph-semantic-icon.tsx|48|12|string-literal|Constraint
 src/features/flowchart/components/graph-semantic-icon.tsx|58|12|string-literal|Cron workstation
@@ -208,5 +208,8 @@ src/features/work-outcome/lib/trends.ts|212|12|string-literal|Tick
 src/features/work-outcome/lib/trends.ts|254|33|string-literal|Session
 src/features/work-outcome/lib/trends.ts|299|14|string-literal|Work type:
 src/features/workflow-activity/components/mutation-dialog.tsx|75|22|string-literal|fixed inset-0 px-5 py-6
-src/features/workflow-activity/lib/react-flow-current-activity-card-graph.ts|55|10|string-literal|workstation:
-src/features/workflow-activity/lib/react-flow-current-activity-card-graph.ts|59|10|string-literal|place:
+src/features/workflow-activity/lib/react-flow-current-activity-card-draft-edges.ts|12|12|string-literal|workstation:
+src/features/workflow-activity/lib/react-flow-current-activity-card-draft-edges.ts|16|12|string-literal|place:
+src/features/workflow-activity/lib/react-flow-current-activity-card-draft-edges.ts|19|10|string-literal|place: :
+src/features/workflow-activity/lib/react-flow-current-activity-card-graph.ts|68|10|string-literal|workstation:
+src/features/workflow-activity/lib/react-flow-current-activity-card-graph.ts|72|10|string-literal|place:
